refactor(datadog): split audit trail processors into per-resource subdirectories
- Move monitor and restriction-policy audit processors into audit_trails/monitor/
- Move user and role-membership audit processors into audit_trails/user/
- Move SLO and SLO restriction-policy audit processors into audit_trails/slo/
- Register SloRestrictionPolicyWebhookProcessor in the webhook registry
- Add configurable notification rule tags support (monitorNotificationRuleTags)
- Drop 404 fallback try/except blocks from audit trail processors
- Update all test imports and patch paths to reflect new module structure
- Remove obsolete test_handle_single_event_404_returns_deleted test cases
