+++ integrations/datadog/datadog/types.py
@@ -1,0 +1,6 @@
+from typing import NamedTuple
+
+
+class RestrictionPolicyResource(NamedTuple):
+    type: str
+    id: str
