+++ integrations/datadog/datadog/webhook/registry.py
@@ -4,19 +4,28 @@ from datadog.webhook.webhook_client import (
     AUDIT_TRAIL_WEBHOOK_PATH,
     MONITOR_WEBHOOK_PATH,
 )
-from datadog.webhook.webhook_processors.audit_trails.monitor_webhook_processor import (
+from datadog.webhook.webhook_processors.audit_trails.monitor.monitor_webhook_processor import (
     MonitorWebhookProcessor as AuditMonitorWebhookProcessor,
+)
+from datadog.webhook.webhook_processors.audit_trails.monitor.monitor_restriction_policy_webhook_processor import (
+    MonitorRestrictionPolicyWebhookProcessor,
 )
 from datadog.webhook.webhook_processors.audit_trails.role_webhook_processor import (
     RoleWebhookProcessor,
 )
-from datadog.webhook.webhook_processors.audit_trails.slo_webhook_processor import (
+from datadog.webhook.webhook_processors.audit_trails.slo.slo_restriction_policy_webhook_processor import (
+    SloRestrictionPolicyWebhookProcessor,
+)
+from datadog.webhook.webhook_processors.audit_trails.slo.slo_webhook_processor import (
     SloWebhookProcessor,
 )
 from datadog.webhook.webhook_processors.audit_trails.team_webhook_processor import (
     TeamWebhookProcessor,
 )
-from datadog.webhook.webhook_processors.audit_trails.user_webhook_processor import (
+from datadog.webhook.webhook_processors.audit_trails.user.role_membership_webhook_processor import (
+    RoleMembershipWebhookProcessor,
+)
+from datadog.webhook.webhook_processors.audit_trails.user.user_webhook_processor import (
     UserWebhookProcessor,
 )
 from datadog.webhook.webhook_processors.monitor_events.monitor_webhook_processor import (
@@ -31,7 +40,16 @@ def register_live_events_webhooks() -> None:
     ocean.add_webhook_processor(MONITOR_WEBHOOK_PATH, MonitorWebhookProcessor)
     ocean.add_webhook_processor(MONITOR_WEBHOOK_PATH, ServiceDependencyWebhookProcessor)
     ocean.add_webhook_processor(AUDIT_TRAIL_WEBHOOK_PATH, AuditMonitorWebhookProcessor)
+    ocean.add_webhook_processor(
+        AUDIT_TRAIL_WEBHOOK_PATH, MonitorRestrictionPolicyWebhookProcessor
+    )
     ocean.add_webhook_processor(AUDIT_TRAIL_WEBHOOK_PATH, UserWebhookProcessor)
+    ocean.add_webhook_processor(
+        AUDIT_TRAIL_WEBHOOK_PATH, RoleMembershipWebhookProcessor
+    )
     ocean.add_webhook_processor(AUDIT_TRAIL_WEBHOOK_PATH, TeamWebhookProcessor)
     ocean.add_webhook_processor(AUDIT_TRAIL_WEBHOOK_PATH, SloWebhookProcessor)
+    ocean.add_webhook_processor(
+        AUDIT_TRAIL_WEBHOOK_PATH, SloRestrictionPolicyWebhookProcessor
+    )
     ocean.add_webhook_processor(AUDIT_TRAIL_WEBHOOK_PATH, RoleWebhookProcessor)
