+++ integrations/datadog/datadog/webhook/webhook_processors/audit_trails/monitor/monitor_webhook_processor.py
@@ -1,41 +1,43 @@
 from typing import TYPE_CHECKING, Any, cast
 
 from integration import ObjectKind
-from datadog.webhook.consts import (
-    TEAM_ACTIONS,
-    AuditTrailAction,
-    AuditTrailEventName,
-)
 from port_ocean.core.handlers.port_app_config.models import ResourceConfig
 
 if TYPE_CHECKING:
-    from datadog.overrides import TeamResourceConfig
+    from datadog.overrides import MonitorResourceConfig
 from port_ocean.core.handlers.webhook.webhook_event import WebhookEventRawResults
 
-from datadog.core.exporters import TeamExporter
-from datadog.core.exporters.team_exporter import GetTeamOptions
+from datadog.core.exporters import MonitorExporter
+from datadog.core.exporters.monitor_exporter import GetMonitorOptions
+from datadog.webhook.consts import (
+    MONITOR_ACTIONS,
+    AuditTrailAction,
+    AuditTrailAssetType,
+    AuditTrailEventName,
+)
 from datadog.webhook.types import AuditTrailEvent
 from datadog.webhook.webhook_processors.audit_trails.base_processor import (
     BaseAuditTrailProcessor,
 )
 
 
-class TeamWebhookProcessor(BaseAuditTrailProcessor):
+class MonitorWebhookProcessor(BaseAuditTrailProcessor):
     async def get_matching_kinds(self, _: Any) -> list[str]:
-        return [ObjectKind.TEAM]
+        return [ObjectKind.MONITOR]
 
     async def _should_process(self, event: AuditTrailEvent) -> bool:
-        # https://docs.datadoghq.com/account_management/audit_trail/events/#teams-management
+        # https://docs.datadoghq.com/account_management/audit_trail/events/#monitor
         attrs = event.attributes
         return (
-            attrs.evt.name == AuditTrailEventName.TEAMS_MANAGEMENT
-            and attrs.action in TEAM_ACTIONS
+            attrs.evt.name == AuditTrailEventName.MONITOR
+            and attrs.asset.type == AuditTrailAssetType.MONITOR
+            and attrs.action in MONITOR_ACTIONS
         )
 
     async def _handle_audit_event(
         self, event: AuditTrailEvent, resource_config: ResourceConfig
     ) -> WebhookEventRawResults:
-        team_id = event.attributes.asset.id
+        monitor_id = event.attributes.asset.id
 
         if event.attributes.action == AuditTrailAction.DELETED:
             return WebhookEventRawResults(
@@ -43,12 +45,13 @@ class TeamWebhookProcessor(BaseAuditTrailProcessor):
                 deleted_raw_results=[event.attributes.asset.dict()],
             )
 
-        team = await TeamExporter(self.client).get_resource(
-            GetTeamOptions.from_resource_config(
-                cast("TeamResourceConfig", resource_config), id=team_id
+        monitor = await MonitorExporter(self.client).get_resource(
+            GetMonitorOptions.from_resource_config(
+                cast("MonitorResourceConfig", resource_config),
+                id=monitor_id,
             )
         )
 
         return WebhookEventRawResults(
-            updated_raw_results=[team] if team else [], deleted_raw_results=[]
+            updated_raw_results=[monitor] if monitor else [], deleted_raw_results=[]
         )
