+++ integrations/datadog/datadog/webhook/webhook_processors/audit_trails/user/user_webhook_processor.py
@@ -1,47 +1,50 @@
 from typing import Any
 
 from integration import ObjectKind
-from datadog.webhook.types import AuditTrailEvent
 from port_ocean.core.handlers.port_app_config.models import ResourceConfig
 from port_ocean.core.handlers.webhook.webhook_event import WebhookEventRawResults
 
-from datadog.core.exporters import RoleExporter
+from datadog.core.exporters import UserExporter
+from datadog.webhook.consts import (
+    USER_ACTIONS,
+    AuditTrailAction,
+    AuditTrailAssetType,
+    AuditTrailEventName,
+)
+from datadog.webhook.types import AuditTrailEvent
 from datadog.webhook.webhook_processors.audit_trails.base_processor import (
     BaseAuditTrailProcessor,
 )
-from datadog.webhook.consts import (
-    AuditTrailAction,
-    AuditTrailEventName,
-    ROLES_ACTIONS,
-)
 
 
-class RoleWebhookProcessor(BaseAuditTrailProcessor):
+class UserWebhookProcessor(BaseAuditTrailProcessor):
     async def get_matching_kinds(self, _: Any) -> list[str]:
-        return [ObjectKind.ROLE]
+        return [ObjectKind.USER]
 
     async def _should_process(self, event: AuditTrailEvent) -> bool:
         # https://docs.datadoghq.com/account_management/audit_trail/events/#access-management
         attrs = event.attributes
         return (
             attrs.evt.name == AuditTrailEventName.ACCESS_MANAGEMENT
-            and attrs.asset.type == ObjectKind.ROLE
-            and attrs.action in ROLES_ACTIONS
+            and attrs.asset.type == AuditTrailAssetType.USER
+            and attrs.action in USER_ACTIONS
         )
 
     async def _handle_audit_event(
         self, event: AuditTrailEvent, resource_config: ResourceConfig
     ) -> WebhookEventRawResults:
-        role_id = event.attributes.asset.id
+        del resource_config
+        attrs = event.attributes
+        user_id = attrs.asset.id
 
-        if event.attributes.action == AuditTrailAction.DELETED:
+        if attrs.action == AuditTrailAction.DELETED:
             return WebhookEventRawResults(
                 updated_raw_results=[],
-                deleted_raw_results=[event.attributes.asset.dict()],
+                deleted_raw_results=[attrs.asset.dict()],
             )
 
-        role = await RoleExporter(self.client).get_resource(role_id)
+        user = await UserExporter(self.client).get_resource(user_id)
 
         return WebhookEventRawResults(
-            updated_raw_results=[role] if role else [], deleted_raw_results=[]
+            updated_raw_results=[user] if user else [], deleted_raw_results=[]
         )
